services/skill-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blanky-skill-server"
version = "0.1.0"
description = "Local-only FastAPI skill server for blanky"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "google-genai>=1.0.0",
  "httpx>=0.27.0",
  "pillow>=10.0.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "tweepy>=4.14.0",
  "uvicorn>=0.32.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
